+++ packages/ddd/src/shared/domain/index.ts
@@ -1,3 +1,4 @@
+export * from './domain-error.registry';
 export * from './domain.error';
 // common errors
 export * from './errors';